+++ app/src/main/java/xyz/alexschubi/ttimer/itemlist/SwipeItemLeft.kt
@@ -19,7 +19,7 @@ class SwipeItemLeft(var adapter: RecyclerViewAdapter, private var fragmentItemLi
     }
     override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
         //val position = viewHolder.bindingAdapterPosition
-        val itemIndex = viewHolder.itemView.id
+        val itemIndex = viewHolder.itemView.id.toLong()
         fragmentItemList.removeItem(adapter.mItems.find { it.Index == itemIndex }!!)
         Functions().deleteItem(itemIndex)
     }
